pyproject.toml
[project]
name = "ai-market-agent-vaia"
version = "0.1.0"
description = "AI Market Analyst - RAG-based Q&A, Summarization, and Data Extraction"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.0",
    "uvicorn>=0.35.0",
    "openai>=2.6.1",
    "pgvector>=0.4.1",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.0",
    "python-dotenv>=1.0.1",
    "tiktoken>=0.12.0",
]